requirements.txt
numpy==1.26.4
pandas==2.2.2
mysql-connector-python==9.1.0
plotly-express==0.4.1
dash==2.18.0
SQLAlchemy==2.0.36